target/dist/flask_app-1.0.dev0/homepage.py
from datetime import datetime
import pytz

#/convert_epoch?epoch_time=<epoch time>
def convert_epoch_to_datetime(epoch_time):
    if epoch_time is None:
        return '<h1>Error: Epoch time parameter is required.</h1>', 400

    try:
        epoch_time = float(epoch_time)
        if epoch_time < 0:
            raise ValueError("Epoch time cannot be negative.")
    except ValueError as e:
        return f'<h1>Error: {str(e)}</h1>', 400
    result = epoch_to_datetime(epoch_time)
    html_result = f'''
    <html>
    <head>
        <style>
            body {{
                font-family: Arial, sans-serif;
                background-color: #f2f2f2;
            }}
            .container {{
                margin: 50px auto;
                padding: 20px;
                background-color: #fff;
                border-radius: 5px;
                box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
                width: 400px;
            }}
            h1 {{
                color: #333;
            }}
            p {{
                color: #666;
            }}
        </style>
    </head>
    <body>
        <div class="container">
            <h1>Converted Datetime:</h1>
            <p>{result}</p>
        </div>
    </body>
    </html>
    '''

    return html_result


def epoch_to_datetime(epoch_time):
    try:
        # Convert epoch time to utc
        utc_time = datetime.utcfromtimestamp(epoch_time).replace(tzinfo=pytz.utc)
        return utc_time.strftime('%Y-%m-%d %H:%M:%S')
    except Exception as e:
        return str(e)
